Make client ID u128
End game is to use random client IDs.

But there's a lot of places where we assume sequential. This is an
easier transition than immediately using UUIDs.

+++ spectrum/src/cli.rs
@@ -90,7 +90,7 @@ pub struct ExperimentArgs {
     /// This includes both viewers and broadcasters (one per channel), so it
     /// must be at least as great as the number of channels.
     #[clap(long, default_value = "10")]
-    clients: u16,
+    clients: u128,
 
     /// Number of workers per group.
     #[clap(long, default_value = "2")]
